requirements.txt
azureml-sdk
argparse
pandas
scikit-learn
tqdm
numpy==1.23